add features to trss
trss is an rss feed reader. I've added the ability to subscribe to feeds, see the entries therein, and browse the entries via lynx.

One thing left to do is to add functionality for opening the entry url in a web browser.
Other than this, just general error checking is required, and a sweep of the code for beautification and performance.