add pre-authentication to bashrc function gp()
there is a program in randomuser/utils which acts as an
authentication manager for tildegit. it doesn't work right.
when the tildegit password has not been requested recently,
instead of popping up a prompt, it will just return an auth
error. i honestly don't know why this is happening, but doing
this makes gpg-agent aware of the decryption and allows the
password to be accessed by git without user intervention.