add a "solo" option for mlg
I'm on vacation, so it now makes sense for mlg to be solo, instead of
being forevermore plugged into a dock. Now, you can type `disp solo` and
make it configure DP-0 for working, instead of trying to configure
DP-1-1 and DP-1-2.

Additionally, this commit also introduces DISP_ACCESS_IP, to remotely
specify the access IP of `x230t`. This is because the wifi I'm using
right now:

- doesn't have the right subnet
- not the right network prefix
- no static ip assignment
- etc.